feat: 봉사모임 상세 페이지 완성, error: 핸들러 작동오류
