+++ pages/volunteer/recommend/detail/index.js
@@ -1,4 +1,4 @@
-import VolunteerDetail from "../../../../src/components/units/volunteer/detail/VolunteerDetail.container";
+import VolunteerDetail from "../../../../src/components/units/volunteer/recommend/detail/VolunteerDetail.container";
 
 export default function VolunteerDetailPage() {
   return(
