--- src/components/units/volunteer/detail/volunteerDetailHeader/VolunteerDetail.container.js
@@ -1,16 +1,0 @@
-import { useRouter } from "next/router"
-import VolunteerDetailHeaderUI from "./VolunteerDetailHeader.presenter"
-
-export default function VolunteerDetailHeader() {
-  const router = useRouter();
-
-  const PrevPage = () => {
-    router.back();
-  }
-
-  return(
-    <>
-      <VolunteerDetailHeaderUI PrevPage={PrevPage} />
-    </>
-  )
-}
