+++ src/components/units/volunteer/recommend/detail/volunteerDetailHeader/VolunteerDetailHeader.styles.js
@@ -1,0 +1,47 @@
+import styled from "@emotion/styled";
+
+export const WrapperHeader = styled.div`
+  width: 390px;
+  height: 95px;
+  background-color: white;
+  margin: 0 auto;
+`;
+
+export const Header = styled.div`
+  width: 100%;
+  height: 95px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: flex-end;
+  padding-left: 23px;
+  padding-bottom: 1px;
+`;
+
+export const LeftArrowTitleContainer = styled.div`
+  width: 250px;
+  height: 40px;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+  margin-bottom: 5px;
+
+  img {
+    &:hover {
+      cursor: pointer;
+    }
+  }
+`;
+
+export const Title = styled.h1`
+  font-size: 24px;
+  margin-left: 10px;
+`;
+
+export const MenuContainer = styled.div`
+  width: 50px;
+  height: 40px;
+  margin-bottom: 5px;
+  cursor: pointer;
+`;
